Several changes made.
Overlapping changes:
Created consistancy and clarity with  our variable names across all docs. As well as change the names of the food name tag so it can be incremented once we expand for project 4.
Consistancy with the order in which we handle nutrients. Always (P,C,F).
Increment naming conventions for the for loops that pertain to meal, food, and nutrient data to make it easier to follow the nested loops and where info is being stored.

Visual Changes:
I made a few minor format changes to the pages. Such as where the macros are displayed. Using the words "Carb" and "Fat" as opposed to "Carbs" and "Fats".

MealPlan.php and Matrix.js:
MealPlan.php now performs all of the linear algebra functionality of Matrix.js. It sucessfully creates serving array that is used to calculate and print the servings of each food for each meal in a table.

Current Bugs:
If the nutrition values of each food are not extremely polar there is a potential for negative serving sizes.